requirements.txt
pillow
matplotlib
sfsutils
addict
tabulate